refactor: separate navbar comp
