--- quizapp/frontend-blog/src/App css
@@ -1,5 +1,0 @@
-html {
-  body {
-    color: red !important;
-  }
-}
