--- quizapp/frontend-blog/src/App2.css
@@ -1,2 +1,0 @@
-html body {
-  color: red; }
